Testing: Fix temp file error in bin tests
The creation of some tempfiles in `test_bin_rucio` seems to be
problematic. A non-deterministic error occours in the GitHub Actions,
which is triggered by a non-existing temp file.

This PR uses the `make_temp_file` function provided by the `common.py`
in tests to create the temp files, in contrast to the manual creation.