Testing: Refactor the pytest marker definition
The pytest configuration is defined in the `plugin.py` file, not in a
config file. This allows us to flexibly change the configuration
depending on the environment. The pytest _markers_ are defined here as
well.

To not write the boilerplate of adding a configuration to pytest for
every marker, split the data from the implementation. We therefore have
the markers defined in a single place with a fixed schema, which makes
it easier to add new ones.