:biohazard: Annotate a session parameter
The session parameter in combination with the
`{read,stream,transactional}_session` is special. While the default
value for the session is set to `None`, the decorator _ensures_ that the
session is initialized. This way, the session is always initialized.

`Pylance` has some problems with that. Setting the parameter type hint
to `Session` (`SessionMaker` should actually be the valid type here)
does report an error, since the default value is `None`. While `None`
can not be assigned to a `Session` parameter, the decorator ensures that
this never happens.

`Pylance` is a static type checker, let's see if `Pylance` correctly
determines the result...