REST & API: Remove duplicated code in error handling
The `ErrorHandlingMethodView` class only handles generic Rucio
exceptions. Individual exceptions need to be caught by each view
function and handeled seperately. This creates a lot of duplicated code
with no additional information.

Changing the error handling to be central removes the manual addition of
each error that can be thrown and removes duplication.